Move voxel data files to world folder
