+++ src/client/world/chunk_mesh_generation.h
@@ -4,7 +4,7 @@
 
 class Chunk;
 
-#include <common/data/voxel_types.h>
-#include "../data/client_voxel_data.h"
+#include <common/world/voxel_types.h>
+#include "client_voxel.h"
 
-ChunkMesh makeChunkMesh(const Chunk &chunk, const VoxelRegistry<ClientVoxelData>& voxelData);
+ChunkMesh makeChunkMesh(const Chunk &chunk, const VoxelRegistry<ClientVoxel>& voxelData);
